refactor(apps): rename bash compoment
